backend/config/auth_config.py
from jose import JWTError, jwt
from datetime import datetime, timedelta

# Clé secrète à changer en prod
SECRET_KEY = "votre_cle_secrete_super_longue"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
