backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose
passlib[bcrypt]
mysql-connector-python
langchain
langchain-community
langchain-openai
openai
crewai
chromadb
faiss-cpu
tiktoken
unstructured
